split out matcher, update gitignore
