+++ .gitignore
@@ -53,3 +53,4 @@ coverage.xml
 docs/_build/
 
 config.ini
+*.swp
