+++ OrgProbe/signing.py
@@ -11,13 +11,13 @@ else:
 
 class RequestSigner(object):
     def __init__(self, secret):
-        self.secret = secret
+        self.secret = secret.encode('utf8')
 
     def sign(self, *args):
         msg = ':'.join(
             [str(x) if not isinstance(x, str_types) else x for x in args])
         logging.debug("Using signature string: %s", msg)
-        hm = hmac.new(self.secret, msg, hashlib.sha512)
+        hm = hmac.new(self.secret, msg.encode('utf8'), hashlib.sha512)
         return hm.hexdigest()
 
     def get_signature(self, args, keys):
